converted to int

+++ pioneer/BMP180.py
@@ -120,15 +120,15 @@ class BMP180(object):
 
 		# Pressure Calculations
 		B6 = B5 - 4000
-		X1 = (self.cal_B2 * (B6 * B6) >> 12) >> 11
-		X2 = (self.cal_AC2 * B6) >> 11
+		X1 = int(self.cal_B2 * int(B6 * B6) >> 12) >> 11
+		X2 = int(self.cal_AC2 * B6) >> 11
 		X3 = X1 + X2
-		B3 = (((self.cal_AC1 * 4 + X3) << self._mode) + 2) / 4
+		B3 = ((int(self.cal_AC1 * 4 + X3) << self._mode) + 2) / 4
 
-		X1 = (self.cal_AC3 * B6) >> 13
-		X2 = (self.cal_B1 * ((B6 * B6) >> 12)) >> 16
-		X3 = ((X1 + X2) + 2) >> 2
-		B4 = (self.cal_AC4 * (X3 + 32768)) >> 15
+		X1 = int(self.cal_AC3 * B6) >> 13
+		X2 = int(self.cal_B1 * ((B6 * B6) >> 12)) >> 16
+		X3 = int((X1 + X2) + 2) >> 2
+		B4 = int(self.cal_AC4 * (X3 + 32768)) >> 15
 		B7 = (UP - B3) * (50000 >> self._mode)
 
 		if B7 < 0x80000000:
